Flytt mutering av saksinformasjon
